AI-Food/requirement.txt
langchain
langchain-openai
langchain-cohere
langchain-core
python-dotenv
pydantic